requirements.txt
tensorflow>=2.8.0
numpy>=1.21.0
matplotlib>=3.5.0
jupyter>=1.0.0
scikit-learn>=1.0.0
Pillow>=8.3.0